LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/text-anchor-in-vertical-rl-expected.txt

There should be no red below.

FAIL
FAIL
FAIL
FAIL
line

FAIL Line at edge of scrollport shouldn't jump visually when content is inserted before assert_equals: expected -400 but got -300
